# docs/openapi.json
{
  "components": {
    "schemas": {
      "AgentDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/AnalysisSummary"
            },
            "title": "Analyses",
            "type": "array"
          },
          "committee_steps": {
            "items": {
              "$ref": "#/components/schemas/CommitteeStepResponse"
            },
            "title": "Committee Steps",
            "type": "array"
          },
          "decision_architecture": {
            "title": "Decision Architecture",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "model_roster": {
            "$ref": "#/components/schemas/ModelRoster"
          },
          "no_trade_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NoTradeDecisionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "pnl_history": {
            "items": {
              "$ref": "#/components/schemas/PnlHistoryPoint"
            },
            "title": "Pnl History",
            "type": "array"
          },
          "portfolio": {
            "$ref": "#/components/schemas/PortfolioSnapshot"
          },
          "sectors": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Sectors",
            "type": "object"
          },
          "stats": {
            "$ref": "#/components/schemas/AgentStatsResponse"
          },
          "strategy": {
            "$ref": "#/components/schemas/StrategyResponse"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/AgentTradeResponse"
            },
            "title": "Trades",
            "type": "array"
          },
          "user_type": {
            "title": "User Type",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "display_name",
          "user_type",
          "decision_architecture",
          "model_roster",
          "strategy",
          "portfolio",
          "trades",
          "sectors",
          "stats",
          "analyses",
          "committee_steps",
          "no_trade_decision",
          "pnl_history"
        ],
        "title": "AgentDetailResponse",
        "type": "object"
      },
...
      "AgentListResponse": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentListItem"
            },
            "title": "Agents",
            "type": "array"
          }
        },
        "required": [
          "agents"
        ],
        "title": "AgentListResponse",
        "type": "object"
      },
...
      "AnalysisRecord": {
        "additionalProperties": false,
        "properties": {
          "analysis_text": {
            "title": "Analysis Text",
            "type": "string"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "cycle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cycle Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "key_actions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Actions"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "cycle_id",
          "analysis_text",
          "key_actions",
          "confidence_score",
          "created_at",
          "username"
        ],
        "title": "AnalysisRecord",
        "type": "object"
      },
      "AnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "analysis": {
            "title": "Analysis",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "agent",
          "analysis",
          "timestamp"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
      "AnalysisSummary": {
        "additionalProperties": false,
        "properties": {
          "created": {
            "title": "Created",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "created"
        ],
        "title": "AnalysisSummary",
        "type": "object"
      },
      "BuildPortfolioResponse": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "positions": {
            "title": "Positions",
            "type": "integer"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/BuiltPortfolioTrade"
            },
            "title": "Trades",
            "type": "array"
          }
        },
        "required": [
          "agent",
          "positions",
          "trades",
          "timestamp"
        ],
        "title": "BuildPortfolioResponse",
        "type": "object"
      },
...
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "agent",
          "response",
          "timestamp"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "CommitteeStepResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "pi_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pi Session Id"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "response_status": {
            "title": "Response Status",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "usage_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Json"
          }
        },
        "required": [
          "sequence",
          "phase",
          "role",
          "provider",
          "model_name",
          "pi_session_id",
          "usage_json",
          "estimated_cost_usd",
          "response_status",
          "error",
          "created_at"
        ],
        "title": "CommitteeStepResponse",
        "type": "object"
      },
      "CreateAgentRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/StrategyConfig"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "persona": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "style": {
            "default": "balanced",
            "enum": [
              "aggressive",
              "value",
              "balanced"
            ],
            "title": "Style",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "username": {
            "pattern": "^[A-Za-z][A-Za-z0-9_]{1,19}$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "CreateAgentRequest",
        "type": "object"
      },
      "CreateAgentResponse": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/CreatedAgent"
          },
          "ok": {
            "const": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "agent"
        ],
        "title": "CreateAgentResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "TransactionResponse": {
        "additionalProperties": false,
        "properties": {
          "cash_balance_after": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Balance After"
          },
          "cash_balance_before": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Balance Before"
          },
          "decision_snapshot_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Snapshot At"
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "execution_market_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Market State"
          },
          "execution_quote_audit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Quote Audit Id"
          },
          "execution_quote_captured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Quote Captured At"
          },
          "execution_quote_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Quote Source"
          },
          "funnel_cycle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funnel Cycle Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "llm_reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Reasoning"
          },
          "market_closed": {
            "title": "Market Closed",
            "type": "integer"
          },
          "price_per_share": {
            "title": "Price Per Share",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "realized_pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Pnl"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "total_value": {
            "title": "Total Value",
            "type": "number"
          },
          "transaction_type": {
            "title": "Transaction Type",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "user_id",
          "ticker",
          "transaction_type",
          "quantity",
          "price_per_share",
          "total_value",
          "cash_balance_before",
          "cash_balance_after",
          "llm_reasoning",
          "funnel_cycle_id",
          "market_closed",
          "realized_pnl",
          "execution_quote_audit_id",
          "executed_at"
        ],
        "title": "TransactionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Portfolio Simulator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/agent-detail/{username}": {
      "get": {
        "operationId": "agent_detail_api_agent_detail__username__get",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Agent Detail",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents": {
      "get": {
        "operationId": "list_agents_api_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List Agents",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_agent_api_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid operation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/analyses": {
      "get": {
        "operationId": "get_analyses_api_analyses_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRecord"
                  },
                  "title": "Response Get Analyses Api Analyses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Analyses",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/analyze/{agent_name}": {
      "post": {
        "operationId": "deep_analysis_api_analyze__agent_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "title": "Agent Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid operation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Deep Analysis",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/build-portfolio/{agent_name}": {
      "post": {
        "operationId": "build_portfolio_api_build_portfolio__agent_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "title": "Agent Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildPortfolioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid operation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Build Portfolio",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/chat/{agent_name}": {
      "post": {
        "operationId": "chat_with_agent_api_chat__agent_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "title": "Agent Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid operation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Chat With Agent",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/trades/{username}": {
      "get": {
        "description": "Get recent trades for a specific user.",
        "operationId": "user_trades_api_trades__username__get",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response User Trades Api Trades  Username  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "User Trades",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}